% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{create}
\alias{create}
\title{create a container}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#/create-a-container
}
\usage{
create(server, ...)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object}

\item{...}{a named list of parameters}
}
\value{
a \code{\link{docker_container}} object
}
\description{
create a container
}
\examples{
\dontrun{
  create(server, Image = "hello-world")
}

}

