% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{docker_container}
\alias{docker_container}
\title{containers}
\usage{
docker_container(server, id)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object}

\item{id}{the container id}
}
\value{
an object of class \code{docker_container}
}
\description{
containers
}

