% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{docker_image}
\alias{docker_image}
\title{images}
\usage{
docker_image(server, id)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object}

\item{id}{the image id}
}
\value{
an object of class \code{docker_image}
}
\description{
images
}

