% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{docker_list}
\alias{docker_list}
\title{list elements on a server}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#list-containers
}
\usage{
docker_list(server, type = c("containers", "images", "networks", "volumes"),
  ...)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object}

\item{type}{a character}

\item{...}{a named list of parameters}
}
\value{
a tibble
}
\description{
list elements on a server
}

