% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{docker_remove.docker_container}
\alias{docker_remove.docker_container}
\title{remove a container}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#remove-a-container
}
\usage{
\method{docker_remove}{docker_container}(x, ...)
}
\arguments{
\item{x}{a \code{\link{docker_container}} object}

\item{...}{a named list of parameters}
}
\value{
a numeric, the status code of the response
}
\description{
remove a container
}

