% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{docker_rename}
\alias{docker_rename}
\title{rename a container}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#kill-a-container
}
\usage{
docker_rename(container, name)
}
\arguments{
\item{container}{a \code{\link{docker_container}} object}

\item{name}{a character the new container name}
}
\value{
a numeric, the status code of the response
}
\description{
rename a container
}

