% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{docker_restart}
\alias{docker_restart}
\title{restart a container}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#restart-a-container
}
\usage{
docker_restart(container, ...)
}
\arguments{
\item{container}{a \code{\link{docker_container}} object}

\item{...}{a named list of parameters}
}
\value{
a numeric, the status code of the response
}
\description{
restart a container
}

