% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{docker_server}
\alias{docker_server}
\title{server}
\usage{
docker_server(host, port)
}
\arguments{
\item{host}{the IP or URL of the Docker server}

\item{port}{the port of the Docker server}
}
\value{
an object of class \code{docker_server}
}
\description{
server
}

