% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{inspect}
\alias{inspect}
\title{inspect a container}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#inspect-a-container
}
\usage{
inspect(server, id, ...)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object,}

\item{id}{a numeric, the docker container id}

\item{...}{a named list of parameters}
}
\value{
a tibble containing information about this container
}
\description{
inspect a container
}
\keyword{internal}

