% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{list_containers}
\alias{list_containers}
\title{list containers on a server}
\source{
https://docs.docker.com/engine/reference/api/docker_remote_api_v1.24/#list-containers
}
\usage{
list_containers(server, ...)
}
\arguments{
\item{server}{a \code{\link{docker_server}} object}

\item{...}{a named list of parameters}
}
\value{
a tibble
}
\description{
list containers on a server
}

